Functions available in Arrow dplyr queries

The arrow package contains methods for 32 dplyr table functions, many of which are "verbs" that do transformations to one or more tables. The package also has mappings of 205 R functions to the corresponding functions in the Arrow compute library. These allow you to write code inside of dplyr methods that call R functions, including many in packages like stringr and lubridate, and they will get translated to Arrow and run on the Arrow query engine (Acero). This document lists all of the mapped functions.

dplyr verbs

Most verb functions return an arrow_dplyr_query object, similar in spirit to a dbplyr::tbl_lazy. This means that the verbs do not eagerly evaluate the query on the data. To run the query, call either compute(), which returns an arrow Table, or collect(), which pulls the resulting Table into an R data.frame.

anti_join()

arrange()

collapse()

collect()

compute()

count()

distinct()

explain()

filter()

full_join()

glimpse()

group_by()

group_by_drop_default()

group_vars()

groups()

inner_join()

left_join()

mutate()

pull()

relocate()

rename()

rename_with()

right_join()

select()

semi_join()

show_query()

summarise()

tally()

transmute()

ungroup()

union()

union_all()

Function mappings

In the list below, any differences in behavior or support between Acero and the R function are listed. If no notes follow the function name, then you can assume that the function works in Acero just as it does in R.

Functions can be called either as pkg::fun() or just fun(), i.e. both str_sub() and stringr::str_sub() work.

In addition to these functions, you can call any of Arrow's 243 compute functions directly. Arrow has many functions that don't map to an existing R function. In other cases where there is an R function mapping, you can still call the Arrow function directly if you don't want the adaptations that the R mapping has that make Acero behave like R. These functions are listed in the C++ documentation, and in the function registry in R, they are named with an arrow_ prefix, such as arrow_ascii_is_decimal.

arrow

add_filename()

cast()

base

-

!

!=

*

/

&

%/%

%%

%in%

^

+

<

<=

==

>

>=

|

abs()

acos()

all()

any()

as.character()

as.Date()

as.difftime()

as.double()

as.integer()

as.logical()

as.numeric()

asin()

ceiling()

cos()

data.frame()

difftime()

endsWith()

exp()

floor()

format()

grepl()

gsub()

ifelse()

is.character()

is.double()

is.factor()

is.finite()

is.infinite()

is.integer()

is.list()

is.logical()

is.na()

is.nan()

is.numeric()

ISOdate()

ISOdatetime()

log()

log10()

log1p()

log2()

logb()

max()

mean()

min()

nchar()

paste(): the collapse argument is not yet supported

paste0(): the collapse argument is not yet supported

pmax()

pmin()

round()

sign()

sin()

sqrt()

startsWith()

strftime()

strptime()

strrep()

strsplit()

sub()

substr()

substring()

sum()

tan()

tolower()

toupper()

trunc()

bit64

as.integer64()

is.integer64()

dplyr

across(): not yet supported inside filter(); purrr-style lambda functions and use of where() selection helper not yet supported

between()

case_when()

coalesce()

desc()

if_else()

n()

n_distinct()

lubridate

am()

as_date()

as_datetime()

ceiling_date()

date()

date_decimal()

day()

ddays()

decimal_date()

dhours()

dmicroseconds()

dmilliseconds()

dminutes()

dmonths()

dmy()

dmy_h()

dmy_hm()

dmy_hms()

dnanoseconds()

dpicoseconds()

dseconds()

dst()

dweeks()

dyears()

dym()

epiweek()

epiyear()

fast_strptime()

floor_date()

format_ISO8601()

hour()

is.Date()

is.instant()

is.POSIXct()

is.timepoint()

isoweek()

isoyear()

leap_year()

make_date()

make_datetime()

make_difftime()

mday()

mdy()

mdy_h()

mdy_hm()

mdy_hms()

minute()

month()

my()

myd()

parse_date_time()

pm()

qday()

quarter()

round_date()

second()

semester()

tz()

wday()

week()

yday()

ydm()

ydm_h()

ydm_hm()

ydm_hms()

year()

ym()

ymd()

ymd_h()

ymd_hm()

ymd_hms()

yq()

methods

is()

rlang

is_character()

is_double()

is_integer()

is_list()

is_logical()

stats

median()

quantile()

sd()

var()

stringi

stri_reverse()

stringr

str_c(): the collapse argument is not yet supported

str_count()

str_detect()

str_dup()

str_ends()

str_length()

str_like(): not yet in a released version of stringr, but it is supported in arrow

str_pad()

str_replace()

str_replace_all()

str_split()

str_starts()

str_sub()

str_to_lower()

str_to_title()

str_to_upper()

str_trim()

tibble

tibble()

tidyselect

all_of()

contains()

ends_with()

everything()

last_col()

matches()

num_range()

one_of()

starts_with()