arrow: Integration to 'Apache' 'Arrow'

'Apache' 'Arrow' https://arrow.apache.org/ is a cross-language development platform for in-memory data. It specifies a standardized language-independent columnar memory format for flat and hierarchical data, organized for efficient analytic operations on modern hardware. This package provides an interface to the 'Arrow C++' library.

See also

Author

Maintainer: Jonathan Keane jkeane@gmail.com

Authors:

Neal Richardson neal.p.richardson@gmail.com

Ian Cook ianmcook@gmail.com

Nic Crane thisisnic@gmail.com

Dewey Dunnington dewey@fishandwhistle.net (ORCID)

Romain François (ORCID)

Dragoș Moldovan-Grünfeld dragos.mold@gmail.com

Jeroen Ooms jeroen@berkeley.edu

Jacob Wujciak-Jens jacob@wujciak.de

Apache Arrow dev@arrow.apache.org [copyright holder]

Other contributors:

Javier Luraschi javier@rstudio.com [contributor]

Karl Dunkle Werner karldw@users.noreply.github.com (ORCID) [contributor]

Jeffrey Wong jeffreyw@netflix.com [contributor]