Call an Arrow compute function

This function provides a lower-level API for calling Arrow functions by their string function name. You won't use it directly for most applications. Many Arrow compute functions are mapped to R methods, and in a dplyr evaluation context, all Arrow functions are callable with an arrow_ prefix.

call_function(
  function_name,
  ...,
  args = list(...),
  options = empty_named_list()
)

Arguments

function_name
string Arrow compute function name

...
Function arguments, which may include Array, ChunkedArray, Scalar, RecordBatch, or Table.

args
list arguments as an alternative to specifying in ...

options
named list of C++ function options.

Value

An Array, ChunkedArray, Scalar, RecordBatch, or Table, whatever the compute function results in.

See also

Examples

# \donttest{
a <- Array$create(c(1L, 2L, 3L, NA, 5L))
s <- Scalar$create(4L)
call_function("fill_null", a, s)

#> Array
#> <int32>
#> [
#>   1,
#>   2,
#>   3,
#>   4,
#>   5
#> ]

a <- Array$create(rnorm(10000))
call_function("quantile", a, options = list(q = seq(0, 1, 0.25)))

#> Array
#> <double>
#> [
#>   -3.30418,
#>   -0.675502,
#>   0.0011219,
#>   0.674598,
#>   3.58895
#> ]

# }