Apache Arrow data types

These functions create type objects corresponding to Arrow types. Use them when defining a schema() or as inputs to other types, like struct. Most of these functions don't take arguments, but a few do.

int8()

int16()

int32()

int64()

uint8()

uint16()

uint32()

uint64()

float16()

halffloat()

float32()

float()

float64()

boolean()

bool()

utf8()

binary(byte_width = NULL)

string()

date32()

date64()

time32(unit = c("ms", "s"))

time64(unit = c("ns", "us"))

null()

timestamp(unit = c("s", "ms", "us", "ns"), timezone = "")

decimal(precision, scale)

list_of(type)

struct(...)

Arguments

byte_width
For binary(), an optional integer width to create a FixedSizeBinary type. The default NULL results in a BinaryType with variable width.

unit
For time/timestamp types, the time unit. time32() can take either "s" or "ms", while time64() can be "us" or "ns". timestamp() can take any of those four values.

timezone
For timestamp(), an optional time zone string.

precision
For decimal(), precision

scale
For decimal(), scale

type
For list_of(), a data type to make a list-of-type

...
For struct(), a named list of types to define the struct columns

Value

An Arrow type object inheriting from DataType.

Details

A few functions have aliases:

utf8() and string()

float16() and halffloat()

float32() and float()

bool() and boolean()

Called from schema() or struct(), double() also is supported as a way of creating a float64()

date32() creates a datetime type with a "day" unit, like the R Date class. date64() has a "ms" unit.

See also

Examples

# \donttest{
bool()

#> Boolean
#> bool

struct(a = int32(), b = double())

#> StructType
#> struct<a: int32, b: double>

timestamp("ms", timezone = "CEST")

#> Timestamp
#> timestamp[ms, tz=CEST]

time64("ns")

#> Time64
#> time64[ns]

# }