List available Arrow C++ compute functions

This function lists the names of all available Arrow C++ library compute functions. These can be called by passing to call_function(), or they can be called by name with an arrow_ prefix inside a dplyr verb.

list_compute_functions(pattern = NULL, ...)

Arguments

pattern
Optional regular expression to filter the function list

...
Additional parameters passed to grep()

Value

A character vector of available Arrow C++ function names

Details

The resulting list describes the capabilities of your arrow build. Some functions, such as string and regular expression functions, require optional build-time C++ dependencies. If your arrow package was not compiled with those features enabled, those functions will not appear in this list.

Some functions take options that need to be passed when calling them (in a list called options). These options require custom handling in C++; many functions already have that handling set up but not all do. If you encounter one that needs special handling for options, please report an issue.

Note that this list does not enumerate all of the R bindings for these functions. The package includes Arrow methods for many base R functions that can be called directly on Arrow objects, as well as some tidyverse-flavored versions available inside dplyr verbs.

Examples

list_compute_functions() 

#>   [1] "abs"                      "abs_checked"             
#>   [3] "acos"                     "acos_checked"            
#>   [5] "add"                      "add_checked"             
#>   [7] "all"                      "and"                     
#>   [9] "and_kleene"               "and_not"                 
#>  [11] "and_not_kleene"           "any"                     
#>  [13] "array_filter"             "array_sort_indices"      
#>  [15] "array_take"               "ascii_center"            
#>  [17] "ascii_is_alnum"           "ascii_is_alpha"          
#>  [19] "ascii_is_decimal"         "ascii_is_lower"          
#>  [21] "ascii_is_printable"       "ascii_is_space"          
#>  [23] "ascii_is_title"           "ascii_is_upper"          
#>  [25] "ascii_lower"              "ascii_lpad"              
#>  [27] "ascii_ltrim"              "ascii_ltrim_whitespace"  
#>  [29] "ascii_reverse"            "ascii_rpad"              
#>  [31] "ascii_rtrim"              "ascii_rtrim_whitespace"  
#>  [33] "ascii_split_whitespace"   "ascii_trim"              
#>  [35] "ascii_trim_whitespace"    "ascii_upper"             
#>  [37] "asin"                     "asin_checked"            
#>  [39] "atan"                     "atan2"                   
#>  [41] "binary_join"              "binary_join_element_wise"
#>  [43] "binary_length"            "binary_replace_slice"    
#>  [45] "bit_wise_and"             "bit_wise_not"            
#>  [47] "bit_wise_or"              "bit_wise_xor"            
#>  [49] "case_when"                "cast"                    
#>  [51] "ceil"                     "coalesce"                
#>  [53] "cos"                      "cos_checked"             
#>  [55] "count"                    "count_substring"         
#>  [57] "count_substring_regex"    "day"                     
#>  [59] "day_of_week"              "day_of_year"             
#>  [61] "dictionary_encode"        "divide"                  
#>  [63] "divide_checked"           "ends_with"               
#>  [65] "equal"                    "extract_regex"           
#>  [67] "fill_null"                "filter"                  
#>  [69] "find_substring"           "find_substring_regex"    
#>  [71] "floor"                    "greater"                 
#>  [73] "greater_equal"            "hash_count"              
#>  [75] "hash_min_max"             "hash_sum"                
#>  [77] "hour"                     "if_else"                 
#>  [79] "index"                    "index_in"                
#>  [81] "index_in_meta_binary"     "invert"                  
#>  [83] "is_finite"                "is_in"                   
#>  [85] "is_in_meta_binary"        "is_inf"                  
#>  [87] "is_nan"                   "is_null"                 
#>  [89] "is_valid"                 "iso_calendar"            
#>  [91] "iso_week"                 "iso_year"                
#>  [93] "less"                     "less_equal"              
#>  [95] "list_flatten"             "list_parent_indices"     
#>  [97] "list_value_length"        "ln"                      
#>  [99] "ln_checked"               "log10"                   
#> [101] "log10_checked"            "log1p"                   
#> [103] "log1p_checked"            "log2"                    
#> [105] "log2_checked"             "make_struct"             
#> [107] "match_like"               "match_substring"         
#> [109] "match_substring_regex"    "max_element_wise"        
#> [111] "mean"                     "microsecond"             
#> [113] "millisecond"              "min_element_wise"        
#> [115] "min_max"                  "minute"                  
#> [117] "mode"                     "month"                   
#> [119] "multiply"                 "multiply_checked"        
#> [121] "nanosecond"               "negate"                  
#> [123] "negate_checked"           "not_equal"               
#> [125] "or"                       "or_kleene"               
#> [127] "partition_nth_indices"    "power"                   
#> [129] "power_checked"            "quantile"                
#> [131] "quarter"                  "replace_substring"       
#> [133] "replace_substring_regex"  "replace_with_mask"       
#> [135] "second"                   "shift_left"              
#> [137] "shift_left_checked"       "shift_right"             
#> [139] "shift_right_checked"      "sign"                    
#> [141] "sin"                      "sin_checked"             
#> [143] "sort_indices"             "split_pattern"           
#> [145] "split_pattern_regex"      "starts_with"             
#> [147] "stddev"                   "string_is_ascii"         
#> [149] "strptime"                 "subsecond"               
#> [151] "subtract"                 "subtract_checked"        
#> [153] "sum"                      "take"                    
#> [155] "tan"                      "tan_checked"             
#> [157] "tdigest"                  "trunc"                   
#> [159] "unique"                   "utf8_center"             
#> [161] "utf8_is_alnum"            "utf8_is_alpha"           
#> [163] "utf8_is_decimal"          "utf8_is_digit"           
#> [165] "utf8_is_lower"            "utf8_is_numeric"         
#> [167] "utf8_is_printable"        "utf8_is_space"           
#> [169] "utf8_is_title"            "utf8_is_upper"           
#> [171] "utf8_length"              "utf8_lower"              
#> [173] "utf8_lpad"                "utf8_ltrim"              
#> [175] "utf8_ltrim_whitespace"    "utf8_replace_slice"      
#> [177] "utf8_reverse"             "utf8_rpad"               
#> [179] "utf8_rtrim"               "utf8_rtrim_whitespace"   
#> [181] "utf8_slice_codeunits"     "utf8_split_whitespace"   
#> [183] "utf8_trim"                "utf8_trim_whitespace"    
#> [185] "utf8_upper"               "value_counts"            
#> [187] "variance"                 "xor"                     
#> [189] "year"                    

list_compute_functions(pattern = "^UTF8", ignore.case = TRUE)

#>  [1] "utf8_center"           "utf8_is_alnum"         "utf8_is_alpha"        
#>  [4] "utf8_is_decimal"       "utf8_is_digit"         "utf8_is_lower"        
#>  [7] "utf8_is_numeric"       "utf8_is_printable"     "utf8_is_space"        
#> [10] "utf8_is_title"         "utf8_is_upper"         "utf8_length"          
#> [13] "utf8_lower"            "utf8_lpad"             "utf8_ltrim"           
#> [16] "utf8_ltrim_whitespace" "utf8_replace_slice"    "utf8_reverse"         
#> [19] "utf8_rpad"             "utf8_rtrim"            "utf8_rtrim_whitespace"
#> [22] "utf8_slice_codeunits"  "utf8_split_whitespace" "utf8_trim"            
#> [25] "utf8_trim_whitespace"  "utf8_upper"           

list_compute_functions(pattern = "^is", invert = TRUE)

#>   [1] "abs"                      "abs_checked"             
#>   [3] "acos"                     "acos_checked"            
#>   [5] "add"                      "add_checked"             
#>   [7] "all"                      "and"                     
#>   [9] "and_kleene"               "and_not"                 
#>  [11] "and_not_kleene"           "any"                     
#>  [13] "array_filter"             "array_sort_indices"      
#>  [15] "array_take"               "ascii_center"            
#>  [17] "ascii_is_alnum"           "ascii_is_alpha"          
#>  [19] "ascii_is_decimal"         "ascii_is_lower"          
#>  [21] "ascii_is_printable"       "ascii_is_space"          
#>  [23] "ascii_is_title"           "ascii_is_upper"          
#>  [25] "ascii_lower"              "ascii_lpad"              
#>  [27] "ascii_ltrim"              "ascii_ltrim_whitespace"  
#>  [29] "ascii_reverse"            "ascii_rpad"              
#>  [31] "ascii_rtrim"              "ascii_rtrim_whitespace"  
#>  [33] "ascii_split_whitespace"   "ascii_trim"              
#>  [35] "ascii_trim_whitespace"    "ascii_upper"             
#>  [37] "asin"                     "asin_checked"            
#>  [39] "atan"                     "atan2"                   
#>  [41] "binary_join"              "binary_join_element_wise"
#>  [43] "binary_length"            "binary_replace_slice"    
#>  [45] "bit_wise_and"             "bit_wise_not"            
#>  [47] "bit_wise_or"              "bit_wise_xor"            
#>  [49] "case_when"                "cast"                    
#>  [51] "ceil"                     "coalesce"                
#>  [53] "cos"                      "cos_checked"             
#>  [55] "count"                    "count_substring"         
#>  [57] "count_substring_regex"    "day"                     
#>  [59] "day_of_week"              "day_of_year"             
#>  [61] "dictionary_encode"        "divide"                  
#>  [63] "divide_checked"           "ends_with"               
#>  [65] "equal"                    "extract_regex"           
#>  [67] "fill_null"                "filter"                  
#>  [69] "find_substring"           "find_substring_regex"    
#>  [71] "floor"                    "greater"                 
#>  [73] "greater_equal"            "hash_count"              
#>  [75] "hash_min_max"             "hash_sum"                
#>  [77] "hour"                     "if_else"                 
#>  [79] "index"                    "index_in"                
#>  [81] "index_in_meta_binary"     "invert"                  
#>  [83] "less"                     "less_equal"              
#>  [85] "list_flatten"             "list_parent_indices"     
#>  [87] "list_value_length"        "ln"                      
#>  [89] "ln_checked"               "log10"                   
#>  [91] "log10_checked"            "log1p"                   
#>  [93] "log1p_checked"            "log2"                    
#>  [95] "log2_checked"             "make_struct"             
#>  [97] "match_like"               "match_substring"         
#>  [99] "match_substring_regex"    "max_element_wise"        
#> [101] "mean"                     "microsecond"             
#> [103] "millisecond"              "min_element_wise"        
#> [105] "min_max"                  "minute"                  
#> [107] "mode"                     "month"                   
#> [109] "multiply"                 "multiply_checked"        
#> [111] "nanosecond"               "negate"                  
#> [113] "negate_checked"           "not_equal"               
#> [115] "or"                       "or_kleene"               
#> [117] "partition_nth_indices"    "power"                   
#> [119] "power_checked"            "quantile"                
#> [121] "quarter"                  "replace_substring"       
#> [123] "replace_substring_regex"  "replace_with_mask"       
#> [125] "second"                   "shift_left"              
#> [127] "shift_left_checked"       "shift_right"             
#> [129] "shift_right_checked"      "sign"                    
#> [131] "sin"                      "sin_checked"             
#> [133] "sort_indices"             "split_pattern"           
#> [135] "split_pattern_regex"      "starts_with"             
#> [137] "stddev"                   "string_is_ascii"         
#> [139] "strptime"                 "subsecond"               
#> [141] "subtract"                 "subtract_checked"        
#> [143] "sum"                      "take"                    
#> [145] "tan"                      "tan_checked"             
#> [147] "tdigest"                  "trunc"                   
#> [149] "unique"                   "utf8_center"             
#> [151] "utf8_is_alnum"            "utf8_is_alpha"           
#> [153] "utf8_is_decimal"          "utf8_is_digit"           
#> [155] "utf8_is_lower"            "utf8_is_numeric"         
#> [157] "utf8_is_printable"        "utf8_is_space"           
#> [159] "utf8_is_title"            "utf8_is_upper"           
#> [161] "utf8_length"              "utf8_lower"              
#> [163] "utf8_lpad"                "utf8_ltrim"              
#> [165] "utf8_ltrim_whitespace"    "utf8_replace_slice"      
#> [167] "utf8_reverse"             "utf8_rpad"               
#> [169] "utf8_rtrim"               "utf8_rtrim_whitespace"   
#> [171] "utf8_slice_codeunits"     "utf8_split_whitespace"   
#> [173] "utf8_trim"                "utf8_trim_whitespace"    
#> [175] "utf8_upper"               "value_counts"            
#> [177] "variance"                 "xor"                     
#> [179] "year"