Open a multi-file dataset

Arrow Datasets allow you to query against data that has been split across multiple files. This sharding of data may indicate partitioning, which can accelerate queries that only touch some partitions (files). Call open_dataset() to point to a directory of data files and return a Dataset, then use dplyr methods to query it.

open_dataset(sources, schema = NULL, partitioning = hive_partition(), ...)

Arguments

sources
Either a string path to a directory containing data files, or a list of SourceFactory objects as created by open_source().

schema
Schema for the dataset. If NULL (the default), the schema will be inferred from the data sources.

partitioning
When sources is a file path, one of

A Schema, in which case the file paths relative to sources will be parsed, and path segments will be matched with the schema fields. For example, schema(year = int16(), month = int8()) would create partitions for file paths like "2019/01/file.parquet", "2019/02/file.parquet", etc.

A character vector that defines the field names corresponding to those path segments (that is, you're providing the names that would correspond to a Schema but the types will be autodetected)

A HivePartitioning or HivePartitioningFactory, as returned by hive_partition() which parses explicit or autodetected fields from Hive-style path segments

NULL for no partitioning

...
additional arguments passed to open_source() when sources is a file path, otherwise ignored.

Value

A Dataset R6 object. Use dplyr methods on it to query the data, or call $NewScan() to construct a query directly.

See also