Read Arrow IPC stream format

Apache Arrow defines two formats for serializing data for interprocess communication (IPC): a "stream" format and a "file" format, known as Feather. read_ipc_stream() and read_feather() read those formats, respectively.

Usage

read_ipc_stream(file, as_data_frame = TRUE, ...)

Arguments

file

A character file name or URI, connection, raw vector, an Arrow input stream, or a FileSystem with path (SubTreeFileSystem). If a file name or URI, an Arrow InputStream will be opened and closed when finished. If an input stream is provided, it will be left open.

as_data_frame

Should the function return a tibble (default) or an Arrow Table?

...

extra parameters passed to read_feather().

Value

A tibble if as_data_frame is TRUE (the default), or an Arrow Table otherwise

See also