Read an arrow::Table from a stream

Read an arrow::Table from a stream

read_table(stream)

read_arrow(stream)

Arguments

stream
stream.

a arrow::RecordBatchFileReader: read an arrow::Table from all the record batches in the reader

a arrow::RecordBatchStreamReader: read an arrow::Table from the remaining record batches in the reader

a string file path: interpret the file as an arrow binary file format, and uses a arrow::RecordBatchFileReader to process it.

a raw vector: read using a arrow::RecordBatchStreamReader

Value

read_table returns an arrow::Table

read_arrow returns a data.frame

Details

The methods using arrow::RecordBatchFileReader and arrow::RecordBatchStreamReader offer the most flexibility. The other methods are for convenience.