Write Arrow formatted data

Write Arrow formatted data

write_arrow(x, stream, ...)

Arguments

x
an arrow::Table, an arrow::RecordBatch or a data.frame

stream
where to serialize to

A arrow::ipc::RecordBatchWriter: the $write() of x is used. The stream is left open. This uses the streaming format or the binary file format depending on the type of the writer.

A string or file path: x is serialized with a arrow::ipc::RecordBatchFileWriter, i.e. using the binary file format.

A raw vector: typically of length zero (its data is ignored, and only used for dispatch). x is serialized using the streaming format, i.e. using the arrow::ipc::RecordBatchStreamWriter

...
extra parameters, currently ignored

write_arrow is a convenience function, the classes arrow::ipc::RecordBatchFileWriter and arrow::ipc::RecordBatchStreamWriter can be used for more flexibility.