Write Arrow data to a raw vector

write_ipc_stream() and write_feather() write data to a sink and return the data (data.frame, RecordBatch, or Table) they were given. This function wraps those so that you can serialize data to a buffer and access that buffer as a raw vector in R.

write_to_raw(x, format = c("stream", "file"))

Arguments

x
data.frame, RecordBatch, or Table

format
one of c("stream", "file"), indicating the IPC format to use

Value

A raw vector containing the bytes of the IPC serialized data.

Examples

# The default format is "stream"
write_to_raw(mtcars)

#>    [1] ff ff ff ff f8 03 00 00 10 00 00 00 00 00 0a 00 0e 00 06 00 05 00 08 00
#>   [25] 0a 00 00 00 00 01 04 00 10 00 00 00 00 00 0a 00 0c 00 00 00 04 00 08 00
#>   [49] 0a 00 00 00 b8 01 00 00 04 00 00 00 01 00 00 00 0c 00 00 00 08 00 0c 00
#>   [73] 04 00 08 00 08 00 00 00 08 00 00 00 0c 00 00 00 01 00 00 00 72 00 00 00
#>   [97] 86 01 00 00 41 0a 33 0a 32 36 32 34 30 30 0a 31 39 37 38 38 38 0a 31 34
#>  [121] 0a 41 4e 53 49 5f 58 33 2e 34 2d 31 39 36 38 0a 35 33 31 0a 32 0a 35 33
#>  [145] 31 0a 31 0a 31 36 0a 31 0a 32 36 32 31 35 33 0a 31 30 0a 64 61 74 61 2e
#>  [169] 66 72 61 6d 65 0a 31 30 32 36 0a 31 0a 32 36 32 31 35 33 0a 35 0a 6e 61
#>  [193] 6d 65 73 0a 31 36 0a 31 0a 32 36 32 31 35 33 0a 35 0a 63 6c 61 73 73 0a
#>  [217] 32 35 34 0a 35 33 31 0a 31 31 0a 32 35 34 0a 32 35 34 0a 32 35 34 0a 32
#>  [241] 35 34 0a 32 35 34 0a 32 35 34 0a 32 35 34 0a 32 35 34 0a 32 35 34 0a 32
#>  [265] 35 34 0a 32 35 34 0a 31 30 32 36 0a 35 31 31 0a 31 36 0a 31 31 0a 32 36
#>  [289] 32 31 35 33 0a 33 0a 6d 70 67 0a 32 36 32 31 35 33 0a 33 0a 63 79 6c 0a
#>  [313] 32 36 32 31 35 33 0a 34 0a 64 69 73 70 0a 32 36 32 31 35 33 0a 32 0a 68
#>  [337] 70 0a 32 36 32 31 35 33 0a 34 0a 64 72 61 74 0a 32 36 32 31 35 33 0a 32
#>  [361] 0a 77 74 0a 32 36 32 31 35 33 0a 34 0a 71 73 65 63 0a 32 36 32 31 35 33
#>  [385] 0a 32 0a 76 73 0a 32 36 32 31 35 33 0a 32 0a 61 6d 0a 32 36 32 31 35 33
#>  [409] 0a 34 0a 67 65 61 72 0a 32 36 32 31 35 33 0a 34 0a 63 61 72 62 0a 32 35
#>  [433] 34 0a 31 30 32 36 0a 35 31 31 0a 31 36 0a 32 0a 32 36 32 31 35 33 0a 31
#>  [457] 30 0a 61 74 74 72 69 62 75 74 65 73 0a 32 36 32 31 35 33 0a 37 0a 63 6f
#>  [481] 6c 75 6d 6e 73 0a 32 35 34 0a 00 00 0b 00 00 00 e0 01 00 00 a4 01 00 00
#>  [505] 74 01 00 00 48 01 00 00 18 01 00 00 ec 00 00 00 bc 00 00 00 90 00 00 00
#>  [529] 64 00 00 00 34 00 00 00 04 00 00 00 5c fe ff ff 00 00 01 03 10 00 00 00
#>  [553] 18 00 00 00 04 00 00 00 00 00 00 00 04 00 00 00 63 61 72 62 00 00 00 00
#>  [577] 4e fe ff ff 00 00 02 00 88 fe ff ff 00 00 01 03 10 00 00 00 18 00 00 00
#>  [601] 04 00 00 00 00 00 00 00 04 00 00 00 67 65 61 72 00 00 00 00 7a fe ff ff
#>  [625] 00 00 02 00 b4 fe ff ff 00 00 01 03 10 00 00 00 14 00 00 00 04 00 00 00
#>  [649] 00 00 00 00 02 00 00 00 61 6d 00 00 a2 fe ff ff 00 00 02 00 dc fe ff ff
#>  [673] 00 00 01 03 10 00 00 00 14 00 00 00 04 00 00 00 00 00 00 00 02 00 00 00
#>  [697] 76 73 00 00 ca fe ff ff 00 00 02 00 04 ff ff ff 00 00 01 03 10 00 00 00
#>  [721] 18 00 00 00 04 00 00 00 00 00 00 00 04 00 00 00 71 73 65 63 00 00 00 00
#>  [745] f6 fe ff ff 00 00 02 00 30 ff ff ff 00 00 01 03 10 00 00 00 14 00 00 00
#>  [769] 04 00 00 00 00 00 00 00 02 00 00 00 77 74 00 00 1e ff ff ff 00 00 02 00
#>  [793] 58 ff ff ff 00 00 01 03 10 00 00 00 18 00 00 00 04 00 00 00 00 00 00 00
#>  [817] 04 00 00 00 64 72 61 74 00 00 00 00 4a ff ff ff 00 00 02 00 84 ff ff ff
#>  [841] 00 00 01 03 10 00 00 00 14 00 00 00 04 00 00 00 00 00 00 00 02 00 00 00
#>  [865] 68 70 00 00 72 ff ff ff 00 00 02 00 ac ff ff ff 00 00 01 03 10 00 00 00
#>  [889] 18 00 00 00 04 00 00 00 00 00 00 00 04 00 00 00 64 69 73 70 00 00 00 00
#>  [913] 9e ff ff ff 00 00 02 00 d8 ff ff ff 00 00 01 03 10 00 00 00 14 00 00 00
#>  [937] 04 00 00 00 00 00 00 00 03 00 00 00 63 79 6c 00 c6 ff ff ff 00 00 02 00
#>  [961] 10 00 14 00 08 00 06 00 07 00 0c 00 00 00 10 00 10 00 00 00 00 00 01 03
#>  [985] 10 00 00 00 1c 00 00 00 04 00 00 00 00 00 00 00 03 00 00 00 6d 70 67 00
#> [1009] 00 00 06 00 08 00 06 00 06 00 00 00 00 00 02 00 ff ff ff ff 68 02 00 00
#> [1033] 14 00 00 00 00 00 00 00 0c 00 16 00 06 00 05 00 08 00 0c 00 0c 00 00 00
#> [1057] 00 03 04 00 18 00 00 00 00 0b 00 00 00 00 00 00 00 00 0a 00 18 00 0c 00
#> [1081] 04 00 08 00 0a 00 00 00 7c 01 00 00 10 00 00 00 20 00 00 00 00 00 00 00
#> [1105] 00 00 00 00 16 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1129] 00 00 00 00 00 00 00 00 00 01 00 00 00 00 00 00 00 01 00 00 00 00 00 00
#> [1153] 00 00 00 00 00 00 00 00 00 01 00 00 00 00 00 00 00 01 00 00 00 00 00 00
#> [1177] 00 02 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 02 00 00 00 00 00 00
#> [1201] 00 01 00 00 00 00 00 00 00 03 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1225] 00 03 00 00 00 00 00 00 00 01 00 00 00 00 00 00 00 04 00 00 00 00 00 00
#> [1249] 00 00 00 00 00 00 00 00 00 04 00 00 00 00 00 00 00 01 00 00 00 00 00 00
#> [1273] 00 05 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 05 00 00 00 00 00 00
#> [1297] 00 01 00 00 00 00 00 00 00 06 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1321] 00 06 00 00 00 00 00 00 00 01 00 00 00 00 00 00 00 07 00 00 00 00 00 00
#> [1345] 00 00 00 00 00 00 00 00 00 07 00 00 00 00 00 00 00 01 00 00 00 00 00 00
#> [1369] 00 08 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 08 00 00 00 00 00 00
#> [1393] 00 01 00 00 00 00 00 00 00 09 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1417] 00 09 00 00 00 00 00 00 00 01 00 00 00 00 00 00 00 0a 00 00 00 00 00 00
#> [1441] 00 00 00 00 00 00 00 00 00 0a 00 00 00 00 00 00 00 01 00 00 00 00 00 00
#> [1465] 00 00 00 00 0b 00 00 00 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1489] 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 20 00 00 00 00 00 00 00
#> [1513] 00 00 00 00 00 00 00 00 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1537] 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 20 00 00 00 00 00 00 00
#> [1561] 00 00 00 00 00 00 00 00 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1585] 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 20 00 00 00 00 00 00 00
#> [1609] 00 00 00 00 00 00 00 00 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1633] 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 35 40
#> [1657] 00 00 00 00 00 00 35 40 cd cc cc cc cc cc 36 40 66 66 66 66 66 66 35 40
#> [1681] 33 33 33 33 33 b3 32 40 9a 99 99 99 99 19 32 40 9a 99 99 99 99 99 2c 40
#> [1705] 66 66 66 66 66 66 38 40 cd cc cc cc cc cc 36 40 33 33 33 33 33 33 33 40
#> [1729] cd cc cc cc cc cc 31 40 66 66 66 66 66 66 30 40 cd cc cc cc cc 4c 31 40
#> [1753] 66 66 66 66 66 66 2e 40 cd cc cc cc cc cc 24 40 cd cc cc cc cc cc 24 40
#> [1777] 66 66 66 66 66 66 2d 40 33 33 33 33 33 33 40 40 66 66 66 66 66 66 3e 40
#> [1801] 33 33 33 33 33 f3 40 40 00 00 00 00 00 80 35 40 00 00 00 00 00 00 2f 40
#> [1825] 66 66 66 66 66 66 2e 40 9a 99 99 99 99 99 2a 40 33 33 33 33 33 33 33 40
#> [1849] cd cc cc cc cc 4c 3b 40 00 00 00 00 00 00 3a 40 66 66 66 66 66 66 3e 40
#> [1873] 9a 99 99 99 99 99 2f 40 33 33 33 33 33 b3 33 40 00 00 00 00 00 00 2e 40
#> [1897] 66 66 66 66 66 66 35 40 00 00 00 00 00 00 18 40 00 00 00 00 00 00 18 40
#> [1921] 00 00 00 00 00 00 10 40 00 00 00 00 00 00 18 40 00 00 00 00 00 00 20 40
#> [1945] 00 00 00 00 00 00 18 40 00 00 00 00 00 00 20 40 00 00 00 00 00 00 10 40
#> [1969] 00 00 00 00 00 00 10 40 00 00 00 00 00 00 18 40 00 00 00 00 00 00 18 40
#> [1993] 00 00 00 00 00 00 20 40 00 00 00 00 00 00 20 40 00 00 00 00 00 00 20 40
#> [2017] 00 00 00 00 00 00 20 40 00 00 00 00 00 00 20 40 00 00 00 00 00 00 20 40
#> [2041] 00 00 00 00 00 00 10 40 00 00 00 00 00 00 10 40 00 00 00 00 00 00 10 40
#> [2065] 00 00 00 00 00 00 10 40 00 00 00 00 00 00 20 40 00 00 00 00 00 00 20 40
#> [2089] 00 00 00 00 00 00 20 40 00 00 00 00 00 00 20 40 00 00 00 00 00 00 10 40
#> [2113] 00 00 00 00 00 00 10 40 00 00 00 00 00 00 10 40 00 00 00 00 00 00 20 40
#> [2137] 00 00 00 00 00 00 18 40 00 00 00 00 00 00 20 40 00 00 00 00 00 00 10 40
#> [2161] 00 00 00 00 00 00 64 40 00 00 00 00 00 00 64 40 00 00 00 00 00 00 5b 40
#> [2185] 00 00 00 00 00 20 70 40 00 00 00 00 00 80 76 40 00 00 00 00 00 20 6c 40
#> [2209] 00 00 00 00 00 80 76 40 66 66 66 66 66 56 62 40 9a 99 99 99 99 99 61 40
#> [2233] 33 33 33 33 33 f3 64 40 33 33 33 33 33 f3 64 40 cd cc cc cc cc 3c 71 40
#> [2257] cd cc cc cc cc 3c 71 40 cd cc cc cc cc 3c 71 40 00 00 00 00 00 80 7d 40
#> [2281] 00 00 00 00 00 c0 7c 40 00 00 00 00 00 80 7b 40 cd cc cc cc cc ac 53 40
#> [2305] cd cc cc cc cc ec 52 40 66 66 66 66 66 c6 51 40 66 66 66 66 66 06 5e 40
#> [2329] 00 00 00 00 00 e0 73 40 00 00 00 00 00 00 73 40 00 00 00 00 00 e0 75 40
#> [2353] 00 00 00 00 00 00 79 40 00 00 00 00 00 c0 53 40 33 33 33 33 33 13 5e 40
#> [2377] 66 66 66 66 66 c6 57 40 00 00 00 00 00 f0 75 40 00 00 00 00 00 20 62 40
#> [2401] 00 00 00 00 00 d0 72 40 00 00 00 00 00 40 5e 40 00 00 00 00 00 80 5b 40
#> [2425] 00 00 00 00 00 80 5b 40 00 00 00 00 00 40 57 40 00 00 00 00 00 80 5b 40
#> [2449] 00 00 00 00 00 e0 65 40 00 00 00 00 00 40 5a 40 00 00 00 00 00 a0 6e 40
#> [2473] 00 00 00 00 00 00 4f 40 00 00 00 00 00 c0 57 40 00 00 00 00 00 c0 5e 40
#> [2497] 00 00 00 00 00 c0 5e 40 00 00 00 00 00 80 66 40 00 00 00 00 00 80 66 40
#> [2521] 00 00 00 00 00 80 66 40 00 00 00 00 00 a0 69 40 00 00 00 00 00 e0 6a 40
#> [2545] 00 00 00 00 00 c0 6c 40 00 00 00 00 00 80 50 40 00 00 00 00 00 00 4a 40
#> [2569] 00 00 00 00 00 40 50 40 00 00 00 00 00 40 58 40 00 00 00 00 00 c0 62 40
#> [2593] 00 00 00 00 00 c0 62 40 00 00 00 00 00 a0 6e 40 00 00 00 00 00 e0 65 40
#> [2617] 00 00 00 00 00 80 50 40 00 00 00 00 00 c0 56 40 00 00 00 00 00 40 5c 40
#> [2641] 00 00 00 00 00 80 70 40 00 00 00 00 00 e0 65 40 00 00 00 00 00 f0 74 40
#> [2665] 00 00 00 00 00 40 5b 40 33 33 33 33 33 33 0f 40 33 33 33 33 33 33 0f 40
#> [2689] cd cc cc cc cc cc 0e 40 a4 70 3d 0a d7 a3 08 40 33 33 33 33 33 33 09 40
#> [2713] 14 ae 47 e1 7a 14 06 40 ae 47 e1 7a 14 ae 09 40 85 eb 51 b8 1e 85 0d 40
#> [2737] 5c 8f c2 f5 28 5c 0f 40 5c 8f c2 f5 28 5c 0f 40 5c 8f c2 f5 28 5c 0f 40
#> [2761] 8f c2 f5 28 5c 8f 08 40 8f c2 f5 28 5c 8f 08 40 8f c2 f5 28 5c 8f 08 40
#> [2785] 71 3d 0a d7 a3 70 07 40 00 00 00 00 00 00 08 40 d7 a3 70 3d 0a d7 09 40
#> [2809] 52 b8 1e 85 eb 51 10 40 b8 1e 85 eb 51 b8 13 40 e1 7a 14 ae 47 e1 10 40
#> [2833] 9a 99 99 99 99 99 0d 40 14 ae 47 e1 7a 14 06 40 33 33 33 33 33 33 09 40
#> [2857] d7 a3 70 3d 0a d7 0d 40 a4 70 3d 0a d7 a3 08 40 52 b8 1e 85 eb 51 10 40
#> [2881] b8 1e 85 eb 51 b8 11 40 29 5c 8f c2 f5 28 0e 40 e1 7a 14 ae 47 e1 10 40
#> [2905] f6 28 5c 8f c2 f5 0c 40 52 b8 1e 85 eb 51 0c 40 71 3d 0a d7 a3 70 10 40
#> [2929] f6 28 5c 8f c2 f5 04 40 00 00 00 00 00 00 07 40 8f c2 f5 28 5c 8f 02 40
#> [2953] b8 1e 85 eb 51 b8 09 40 85 eb 51 b8 1e 85 0b 40 ae 47 e1 7a 14 ae 0b 40
#> [2977] 8f c2 f5 28 5c 8f 0c 40 85 eb 51 b8 1e 85 09 40 33 33 33 33 33 33 09 40
#> [3001] 85 eb 51 b8 1e 85 0b 40 85 eb 51 b8 1e 85 0b 40 48 e1 7a 14 ae 47 10 40
#> [3025] d7 a3 70 3d 0a d7 0d 40 3d 0a d7 a3 70 3d 0e 40 00 00 00 00 00 00 15 40
#> [3049] 19 04 56 0e 2d b2 15 40 e1 7a 14 ae 47 61 15 40 9a 99 99 99 99 99 01 40
#> [3073] d7 a3 70 3d 0a d7 f9 3f 5c 8f c2 f5 28 5c fd 3f b8 1e 85 eb 51 b8 03 40
#> [3097] 29 5c 8f c2 f5 28 0c 40 7b 14 ae 47 e1 7a 0b 40 b8 1e 85 eb 51 b8 0e 40
#> [3121] c3 f5 28 5c 8f c2 0e 40 f6 28 5c 8f c2 f5 fe 3f 1f 85 eb 51 b8 1e 01 40
#> [3145] 68 91 ed 7c 3f 35 f8 3f 5c 8f c2 f5 28 5c 09 40 29 5c 8f c2 f5 28 06 40
#> [3169] 8f c2 f5 28 5c 8f 0c 40 3d 0a d7 a3 70 3d 06 40 f6 28 5c 8f c2 75 30 40
#> [3193] 85 eb 51 b8 1e 05 31 40 5c 8f c2 f5 28 9c 32 40 71 3d 0a d7 a3 70 33 40
#> [3217] 85 eb 51 b8 1e 05 31 40 b8 1e 85 eb 51 38 34 40 ae 47 e1 7a 14 ae 2f 40
#> [3241] 00 00 00 00 00 00 34 40 66 66 66 66 66 e6 36 40 cd cc cc cc cc 4c 32 40
#> [3265] 66 66 66 66 66 e6 32 40 66 66 66 66 66 66 31 40 9a 99 99 99 99 99 31 40
#> [3289] 00 00 00 00 00 00 32 40 7b 14 ae 47 e1 fa 31 40 52 b8 1e 85 eb d1 31 40
#> [3313] ec 51 b8 1e 85 6b 31 40 b8 1e 85 eb 51 78 33 40 85 eb 51 b8 1e 85 32 40
#> [3337] 66 66 66 66 66 e6 33 40 c3 f5 28 5c 8f 02 34 40 1f 85 eb 51 b8 de 30 40
#> [3361] cd cc cc cc cc 4c 31 40 52 b8 1e 85 eb d1 2e 40 cd cc cc cc cc 0c 31 40
#> [3385] 66 66 66 66 66 e6 32 40 33 33 33 33 33 b3 30 40 66 66 66 66 66 e6 30 40
#> [3409] 00 00 00 00 00 00 2d 40 00 00 00 00 00 00 2f 40 33 33 33 33 33 33 2d 40
#> [3433] 9a 99 99 99 99 99 32 40 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [3457] 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 00 00
#> [3481] 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 00 00 00 00 00 00 00 00 f0 3f
#> [3505] 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 f0 3f
#> [3529] 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [3553] 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [3577] 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 f0 3f
#> [3601] 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [3625] 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 f0 3f
#> [3649] 00 00 00 00 00 00 00 00 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 00 00
#> [3673] 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 f0 3f
#> [3697] 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 f0 3f
#> [3721] 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [3745] 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [3769] 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [3793] 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [3817] 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 f0 3f
#> [3841] 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 00 00
#> [3865] 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [3889] 00 00 00 00 00 00 00 00 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 f0 3f
#> [3913] 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 f0 3f
#> [3937] 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 10 40
#> [3961] 00 00 00 00 00 00 10 40 00 00 00 00 00 00 10 40 00 00 00 00 00 00 08 40
#> [3985] 00 00 00 00 00 00 08 40 00 00 00 00 00 00 08 40 00 00 00 00 00 00 08 40
#> [4009] 00 00 00 00 00 00 10 40 00 00 00 00 00 00 10 40 00 00 00 00 00 00 10 40
#> [4033] 00 00 00 00 00 00 10 40 00 00 00 00 00 00 08 40 00 00 00 00 00 00 08 40
#> [4057] 00 00 00 00 00 00 08 40 00 00 00 00 00 00 08 40 00 00 00 00 00 00 08 40
#> [4081] 00 00 00 00 00 00 08 40 00 00 00 00 00 00 10 40 00 00 00 00 00 00 10 40
#> [4105] 00 00 00 00 00 00 10 40 00 00 00 00 00 00 08 40 00 00 00 00 00 00 08 40
#> [4129] 00 00 00 00 00 00 08 40 00 00 00 00 00 00 08 40 00 00 00 00 00 00 08 40
#> [4153] 00 00 00 00 00 00 10 40 00 00 00 00 00 00 14 40 00 00 00 00 00 00 14 40
#> [4177] 00 00 00 00 00 00 14 40 00 00 00 00 00 00 14 40 00 00 00 00 00 00 14 40
#> [4201] 00 00 00 00 00 00 10 40 00 00 00 00 00 00 10 40 00 00 00 00 00 00 10 40
#> [4225] 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 00 40
#> [4249] 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 10 40 00 00 00 00 00 00 00 40
#> [4273] 00 00 00 00 00 00 00 40 00 00 00 00 00 00 10 40 00 00 00 00 00 00 10 40
#> [4297] 00 00 00 00 00 00 08 40 00 00 00 00 00 00 08 40 00 00 00 00 00 00 08 40
#> [4321] 00 00 00 00 00 00 10 40 00 00 00 00 00 00 10 40 00 00 00 00 00 00 10 40
#> [4345] 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 00 40 00 00 00 00 00 00 f0 3f
#> [4369] 00 00 00 00 00 00 f0 3f 00 00 00 00 00 00 00 40 00 00 00 00 00 00 00 40
#> [4393] 00 00 00 00 00 00 10 40 00 00 00 00 00 00 00 40 00 00 00 00 00 00 f0 3f
#> [4417] 00 00 00 00 00 00 00 40 00 00 00 00 00 00 00 40 00 00 00 00 00 00 10 40
#> [4441] 00 00 00 00 00 00 18 40 00 00 00 00 00 00 20 40 00 00 00 00 00 00 00 40
#> [4465] ff ff ff ff 00 00 00 00

write_to_raw(mtcars, format = "file")

#>    [1] 41 52 52 4f 57 31 00 00 ff ff ff ff f8 03 00 00 10 00 00 00 00 00 0a 00
#>   [25] 0e 00 06 00 05 00 08 00 0a 00 00 00 00 01 04 00 10 00 00 00 00 00 0a 00
#>   [49] 0c 00 00 00 04 00 08 00 0a 00 00 00 b8 01 00 00 04 00 00 00 01 00 00 00
#>   [73] 0c 00 00 00 08 00 0c 00 04 00 08 00 08 00 00 00 08 00 00 00 0c 00 00 00
#>   [97] 01 00 00 00 72 00 00 00 86 01 00 00 41 0a 33 0a 32 36 32 34 30 30 0a 31
#>  [121] 39 37 38 38 38 0a 31 34 0a 41 4e 53 49 5f 58 33 2e 34 2d 31 39 36 38 0a
#>  [145] 35 33 31 0a 32 0a 35 33 31 0a 31 0a 31 36 0a 31 0a 32 36 32 31 35 33 0a
#>  [169] 31 30 0a 64 61 74 61 2e 66 72 61 6d 65 0a 31 30 32 36 0a 31 0a 32 36 32
#>  [193] 31 35 33 0a 35 0a 6e 61 6d 65 73 0a 31 36 0a 31 0a 32 36 32 31 35 33 0a
#>  [217] 35 0a 63 6c 61 73 73 0a 32 35 34 0a 35 33 31 0a 31 31 0a 32 35 34 0a 32
#>  [241] 35 34 0a 32 35 34 0a 32 35 34 0a 32 35 34 0a 32 35 34 0a 32 35 34 0a 32
#>  [265] 35 34 0a 32 35 34 0a 32 35 34 0a 32 35 34 0a 31 30 32 36 0a 35 31 31 0a
#>  [289] 31 36 0a 31 31 0a 32 36 32 31 35 33 0a 33 0a 6d 70 67 0a 32 36 32 31 35
#>  [313] 33 0a 33 0a 63 79 6c 0a 32 36 32 31 35 33 0a 34 0a 64 69 73 70 0a 32 36
#>  [337] 32 31 35 33 0a 32 0a 68 70 0a 32 36 32 31 35 33 0a 34 0a 64 72 61 74 0a
#>  [361] 32 36 32 31 35 33 0a 32 0a 77 74 0a 32 36 32 31 35 33 0a 34 0a 71 73 65
#>  [385] 63 0a 32 36 32 31 35 33 0a 32 0a 76 73 0a 32 36 32 31 35 33 0a 32 0a 61
#>  [409] 6d 0a 32 36 32 31 35 33 0a 34 0a 67 65 61 72 0a 32 36 32 31 35 33 0a 34
#>  [433] 0a 63 61 72 62 0a 32 35 34 0a 31 30 32 36 0a 35 31 31 0a 31 36 0a 32 0a
#>  [457] 32 36 32 31 35 33 0a 31 30 0a 61 74 74 72 69 62 75 74 65 73 0a 32 36 32
#>  [481] 31 35 33 0a 37 0a 63 6f 6c 75 6d 6e 73 0a 32 35 34 0a 00 00 0b 00 00 00
#>  [505] e0 01 00 00 a4 01 00 00 74 01 00 00 48 01 00 00 18 01 00 00 ec 00 00 00
#>  [529] bc 00 00 00 90 00 00 00 64 00 00 00 34 00 00 00 04 00 00 00 5c fe ff ff
#>  [553] 00 00 01 03 10 00 00 00 18 00 00 00 04 00 00 00 00 00 00 00 04 00 00 00
#>  [577] 63 61 72 62 00 00 00 00 4e fe ff ff 00 00 02 00 88 fe ff ff 00 00 01 03
#>  [601] 10 00 00 00 18 00 00 00 04 00 00 00 00 00 00 00 04 00 00 00 67 65 61 72
#>  [625] 00 00 00 00 7a fe ff ff 00 00 02 00 b4 fe ff ff 00 00 01 03 10 00 00 00
#>  [649] 14 00 00 00 04 00 00 00 00 00 00 00 02 00 00 00 61 6d 00 00 a2 fe ff ff
#>  [673] 00 00 02 00 dc fe ff ff 00 00 01 03 10 00 00 00 14 00 00 00 04 00 00 00
#>  [697] 00 00 00 00 02 00 00 00 76 73 00 00 ca fe ff ff 00 00 02 00 04 ff ff ff
#>  [721] 00 00 01 03 10 00 00 00 18 00 00 00 04 00 00 00 00 00 00 00 04 00 00 00
#>  [745] 71 73 65 63 00 00 00 00 f6 fe ff ff 00 00 02 00 30 ff ff ff 00 00 01 03
#>  [769] 10 00 00 00 14 00 00 00 04 00 00 00 00 00 00 00 02 00 00 00 77 74 00 00
#>  [793] 1e ff ff ff 00 00 02 00 58 ff ff ff 00 00 01 03 10 00 00 00 18 00 00 00
#>  [817] 04 00 00 00 00 00 00 00 04 00 00 00 64 72 61 74 00 00 00 00 4a ff ff ff
#>  [841] 00 00 02 00 84 ff ff ff 00 00 01 03 10 00 00 00 14 00 00 00 04 00 00 00
#>  [865] 00 00 00 00 02 00 00 00 68 70 00 00 72 ff ff ff 00 00 02 00 ac ff ff ff
#>  [889] 00 00 01 03 10 00 00 00 18 00 00 00 04 00 00 00 00 00 00 00 04 00 00 00
#>  [913] 64 69 73 70 00 00 00 00 9e ff ff ff 00 00 02 00 d8 ff ff ff 00 00 01 03
#>  [937] 10 00 00 00 14 00 00 00 04 00 00 00 00 00 00 00 03 00 00 00 63 79 6c 00
#>  [961] c6 ff ff ff 00 00 02 00 10 00 14 00 08 00 06 00 07 00 0c 00 00 00 10 00
#>  [985] 10 00 00 00 00 00 01 03 10 00 00 00 1c 00 00 00 04 00 00 00 00 00 00 00
#> [1009] 03 00 00 00 6d 70 67 00 00 00 06 00 08 00 06 00 06 00 00 00 00 00 02 00
#> [1033] ff ff ff ff 78 02 00 00 14 00 00 00 00 00 00 00 0c 00 18 00 06 00 05 00
#> [1057] 08 00 0c 00 0c 00 00 00 00 03 04 00 1c 00 00 00 48 06 00 00 00 00 00 00
#> [1081] 00 00 00 00 0c 00 1c 00 10 00 04 00 08 00 0c 00 0c 00 00 00 88 01 00 00
#> [1105] 1c 00 00 00 14 00 00 00 20 00 00 00 00 00 00 00 00 00 00 00 04 00 04 00
#> [1129] 04 00 00 00 16 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1153] 00 00 00 00 00 00 00 00 b3 00 00 00 00 00 00 00 b8 00 00 00 00 00 00 00
#> [1177] 00 00 00 00 00 00 00 00 b8 00 00 00 00 00 00 00 57 00 00 00 00 00 00 00
#> [1201] 10 01 00 00 00 00 00 00 00 00 00 00 00 00 00 00 10 01 00 00 00 00 00 00
#> [1225] b6 00 00 00 00 00 00 00 c8 01 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1249] c8 01 00 00 00 00 00 00 9f 00 00 00 00 00 00 00 68 02 00 00 00 00 00 00
#> [1273] 00 00 00 00 00 00 00 00 68 02 00 00 00 00 00 00 bd 00 00 00 00 00 00 00
#> [1297] 28 03 00 00 00 00 00 00 00 00 00 00 00 00 00 00 28 03 00 00 00 00 00 00
#> [1321] d4 00 00 00 00 00 00 00 00 04 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1345] 00 04 00 00 00 00 00 00 d6 00 00 00 00 00 00 00 d8 04 00 00 00 00 00 00
#> [1369] 00 00 00 00 00 00 00 00 d8 04 00 00 00 00 00 00 51 00 00 00 00 00 00 00
#> [1393] 30 05 00 00 00 00 00 00 00 00 00 00 00 00 00 00 30 05 00 00 00 00 00 00
#> [1417] 3e 00 00 00 00 00 00 00 70 05 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1441] 70 05 00 00 00 00 00 00 57 00 00 00 00 00 00 00 c8 05 00 00 00 00 00 00
#> [1465] 00 00 00 00 00 00 00 00 c8 05 00 00 00 00 00 00 79 00 00 00 00 00 00 00
#> [1489] 00 00 00 00 0b 00 00 00 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1513] 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 20 00 00 00 00 00 00 00
#> [1537] 00 00 00 00 00 00 00 00 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1561] 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 20 00 00 00 00 00 00 00
#> [1585] 00 00 00 00 00 00 00 00 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1609] 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 20 00 00 00 00 00 00 00
#> [1633] 00 00 00 00 00 00 00 00 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [1657] 20 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 01 00 00 00 00 00 00
#> [1681] 04 22 4d 18 60 40 82 9c 00 00 00 11 00 01 00 21 35 40 07 00 50 00 35 40
#> [1705] cd cc 01 00 31 36 40 66 01 00 30 35 40 33 01 00 93 b3 32 40 9a 99 99 99
#> [1729] 99 19 08 00 23 99 2c 20 00 15 38 30 00 00 27 00 33 33 33 33 10 00 13 31
#> [1753] 20 00 12 30 10 00 14 4c 10 00 12 2e 10 00 25 cc 24 08 00 01 67 00 22 66
#> [1777] 2d 68 00 23 33 40 38 00 12 3e 10 00 22 f3 40 98 00 14 80 a0 00 13 2f 20
#> [1801] 00 12 2e 90 00 22 99 2a 28 00 04 78 00 22 4c 3b 30 00 23 00 3a 28 00 13
#> [1825] 3e 28 00 12 2f 28 00 23 b3 33 20 00 a0 2e 40 66 66 66 66 66 66 35 40 00
#> [1849] 00 00 00 00 00 00 00 00 00 01 00 00 00 00 00 00 04 22 4d 18 60 40 82 40
#> [1873] 00 00 00 11 00 01 00 21 18 40 07 00 05 08 00 1b 10 10 00 1f 20 10 00 04
#> [1897] 04 28 00 04 08 00 04 40 00 04 08 00 04 38 00 0f 08 00 15 04 48 00 0f 08
#> [1921] 00 05 0f 48 00 0d 0f 38 00 0d 09 98 00 50 00 00 00 10 40 00 00 00 00 00
#> [1945] 00 01 00 00 00 00 00 00 04 22 4d 18 60 40 82 9f 00 00 00 11 00 01 00 21
#> [1969] 64 40 07 00 05 08 00 12 5b 10 00 22 20 70 08 00 22 80 76 08 00 22 20 6c
#> [1993] 08 00 40 80 76 40 66 01 00 50 56 62 40 9a 99 01 00 30 61 40 33 01 00 25
#> [2017] f3 64 08 00 7d cd cc cc cc cc 3c 71 08 00 01 68 00 22 80 7d 48 00 22 c0
#> [2041] 7c 08 00 22 80 7b 28 00 22 ac 53 08 00 22 ec 52 60 00 22 c6 51 08 00 22
#> [2065] 06 5e 28 00 22 e0 73 08 00 13 00 08 00 23 e0 75 10 00 12 79 08 00 22 c0
#> [2089] 53 80 00 22 13 5e 38 00 22 c6 57 18 00 13 f0 28 00 22 20 62 10 00 b0 d0
#> [2113] 72 40 00 00 00 00 00 40 5e 40 00 00 00 00 00 00 00 01 00 00 00 00 00 00
#> [2137] 04 22 4d 18 60 40 82 88 00 00 00 10 00 01 00 30 80 5b 40 07 00 05 08 00
#> [2161] 2a 40 57 10 00 22 e0 65 10 00 22 40 5a 08 00 22 a0 6e 08 00 22 00 4f 08
#> [2185] 00 13 c0 30 00 23 c0 5e 10 00 03 08 00 22 80 66 10 00 0c 08 00 22 a0 69
#> [2209] 18 00 22 e0 6a 08 00 22 c0 6c 08 00 22 80 50 08 00 22 00 4a 08 00 13 40
#> [2233] 10 00 22 40 58 10 00 2a c0 62 08 00 04 88 00 04 a0 00 04 40 00 22 c0 56
#> [2257] 28 00 22 40 5c 08 00 22 80 70 08 00 04 28 00 b0 f0 74 40 00 00 00 00 00
#> [2281] 40 5b 40 00 00 00 00 00 00 01 00 00 00 00 00 00 04 22 4d 18 60 40 82 a6
#> [2305] 00 00 00 11 33 01 00 21 0f 40 07 00 50 33 0f 40 cd cc 01 00 93 0e 40 a4
#> [2329] 70 3d 0a d7 a3 08 18 00 a1 09 40 14 ae 47 e1 7a 14 06 40 07 00 fd 03 ae
#> [2353] 09 40 85 eb 51 b8 1e 85 0d 40 5c 8f c2 f5 28 5c 0f 08 00 01 17 00 2d 8f
#> [2377] 08 08 00 10 71 57 00 41 70 07 40 00 01 00 40 08 40 d7 a3 69 00 a1 09 40
#> [2401] 52 b8 1e 85 eb 51 10 40 07 00 50 b8 13 40 e1 7a 72 00 40 10 40 9a 99 01
#> [2425] 00 15 0d 80 00 02 a8 00 13 09 38 00 15 0d a8 00 0a 40 00 31 11 40 29 99
#> [2449] 00 15 0e 48 00 20 f6 28 11 00 13 0c 68 00 a0 0c 40 71 3d 0a d7 a3 70 10
#> [2473] 40 00 00 00 00 00 00 00 00 01 00 00 00 00 00 00 04 22 4d 18 60 40 82 bd
#> [2497] 00 00 00 91 f6 28 5c 8f c2 f5 04 40 00 01 00 f0 03 07 40 8f c2 f5 28 5c
#> [2521] 8f 02 40 b8 1e 85 eb 51 b8 09 40 06 00 b3 1e 85 0b 40 ae 47 e1 7a 14 ae
#> [2545] 0b 20 00 13 0c 18 00 31 09 40 33 01 00 06 28 00 04 30 00 10 48 2f 00 d0
#> [2569] 47 10 40 d7 a3 70 3d 0a d7 0d 40 3d 0a 0a 00 13 0e 68 00 a1 15 40 19 04
#> [2593] 56 0e 2d b2 15 40 27 00 50 61 15 40 9a 99 01 00 13 01 30 00 20 f9 3f 96
#> [2617] 00 42 28 5c fd 3f 88 00 31 03 40 29 11 00 40 0c 40 7b 14 8a 00 13 0b a0
#> [2641] 00 30 0e 40 c3 af 00 32 c2 0e 40 c8 00 31 fe 3f 1f 81 00 93 01 40 68 91
#> [2665] ed 7c 3f 35 f8 48 00 13 09 40 00 15 06 c0 00 80 3d 0a d7 a3 70 3d 06 40
#> [2689] 00 00 00 00 00 00 00 00 00 01 00 00 00 00 00 00 04 22 4d 18 60 40 82 bf
#> [2713] 00 00 00 f5 10 f6 28 5c 8f c2 75 30 40 85 eb 51 b8 1e 05 31 40 5c 8f c2
#> [2737] f5 28 9c 32 40 71 3d 0a d7 a3 70 33 18 00 f1 02 b8 1e 85 eb 51 38 34 40
#> [2761] ae 47 e1 7a 14 ae 2f 40 00 01 00 30 34 40 66 01 00 a3 e6 36 40 cd cc cc
#> [2785] cc cc 4c 32 10 00 03 08 00 50 66 31 40 9a 99 01 00 13 31 30 00 b0 32 40
#> [2809] 7b 14 ae 47 e1 fa 31 40 52 51 00 93 d1 31 40 ec 51 b8 1e 85 6b 60 00 13
#> [2833] 78 70 00 13 85 40 00 c0 e6 33 40 c3 f5 28 5c 8f 02 34 40 1f a1 00 23 de
#> [2857] 30 68 00 04 40 00 12 2e 10 00 25 0c 31 78 00 10 33 01 00 23 b3 30 10 00
#> [2881] 13 30 78 00 13 2d 08 00 20 2f 40 1f 00 c0 33 33 2d 40 9a 99 99 99 99 99
#> [2905] 32 40 00 00 00 00 00 00 00 01 00 00 00 00 00 00 04 22 4d 18 60 40 82 3a
#> [2929] 00 00 00 1f 00 01 00 02 22 f0 3f 17 00 04 08 00 04 02 00 0f 10 00 05 04
#> [2953] 18 00 0c 08 00 0f 02 00 1d 0c 40 00 0c 10 00 0f 02 00 0d 04 30 00 04 02
#> [2977] 00 0c 10 00 09 02 00 50 00 00 00 f0 3f 00 00 00 00 00 00 00 00 00 00 00
#> [3001] 00 01 00 00 00 00 00 00 04 22 4d 18 60 40 82 27 00 00 00 11 00 01 00 21
#> [3025] f0 3f 07 00 0d 08 00 0f 02 00 5d 0c 80 00 04 10 00 0f 02 00 15 04 30 00
#> [3049] 0f 08 00 12 50 00 00 00 f0 3f 00 00 00 00 00 00 00 01 00 00 00 00 00 00
#> [3073] 04 22 4d 18 60 40 82 40 00 00 00 11 00 01 00 21 10 40 07 00 0d 08 00 13
#> [3097] 08 18 00 0f 08 00 05 0c 30 00 0c 10 00 0f 38 00 05 0f 18 00 05 0c 40 00
#> [3121] 04 10 00 0f 30 00 05 0c 18 00 04 30 00 13 14 b8 00 0f 08 00 0a 50 00 00
#> [3145] 00 10 40 00 00 00 00 00 00 01 00 00 00 00 00 00 04 22 4d 18 60 40 82 62
#> [3169] 00 00 00 11 00 01 00 21 10 40 07 00 05 08 00 22 f0 3f 10 00 04 08 00 13
#> [3193] 00 20 00 04 10 00 04 28 00 04 18 00 04 08 00 04 18 00 04 08 00 13 08 38
#> [3217] 00 0c 08 00 04 20 00 0c 08 00 04 60 00 04 50 00 04 10 00 04 08 00 04 18
#> [3241] 00 04 08 00 04 40 00 04 10 00 0f 28 00 0d 13 18 90 00 a0 20 40 00 00 00
#> [3265] 00 00 00 00 40 00 00 00 00 00 00 00 00 00 00 00 ff ff ff ff 00 00 00 00
#> [3289] 10 00 00 00 0c 00 14 00 06 00 08 00 0c 00 10 00 0c 00 00 00 00 00 04 00
#> [3313] 40 00 00 00 28 00 00 00 04 00 00 00 01 00 00 00 08 04 00 00 00 00 00 00
#> [3337] 80 02 00 00 00 00 00 00 48 06 00 00 00 00 00 00 00 00 00 00 00 00 00 00
#> [3361] 00 00 00 00 00 00 0a 00 0c 00 00 00 04 00 08 00 0a 00 00 00 b8 01 00 00
#> [3385] 04 00 00 00 01 00 00 00 0c 00 00 00 08 00 0c 00 04 00 08 00 08 00 00 00
#> [3409] 08 00 00 00 0c 00 00 00 01 00 00 00 72 00 00 00 86 01 00 00 41 0a 33 0a
#> [3433] 32 36 32 34 30 30 0a 31 39 37 38 38 38 0a 31 34 0a 41 4e 53 49 5f 58 33
#> [3457] 2e 34 2d 31 39 36 38 0a 35 33 31 0a 32 0a 35 33 31 0a 31 0a 31 36 0a 31
#> [3481] 0a 32 36 32 31 35 33 0a 31 30 0a 64 61 74 61 2e 66 72 61 6d 65 0a 31 30
#> [3505] 32 36 0a 31 0a 32 36 32 31 35 33 0a 35 0a 6e 61 6d 65 73 0a 31 36 0a 31
#> [3529] 0a 32 36 32 31 35 33 0a 35 0a 63 6c 61 73 73 0a 32 35 34 0a 35 33 31 0a
#> [3553] 31 31 0a 32 35 34 0a 32 35 34 0a 32 35 34 0a 32 35 34 0a 32 35 34 0a 32
#> [3577] 35 34 0a 32 35 34 0a 32 35 34 0a 32 35 34 0a 32 35 34 0a 32 35 34 0a 31
#> [3601] 30 32 36 0a 35 31 31 0a 31 36 0a 31 31 0a 32 36 32 31 35 33 0a 33 0a 6d
#> [3625] 70 67 0a 32 36 32 31 35 33 0a 33 0a 63 79 6c 0a 32 36 32 31 35 33 0a 34
#> [3649] 0a 64 69 73 70 0a 32 36 32 31 35 33 0a 32 0a 68 70 0a 32 36 32 31 35 33
#> [3673] 0a 34 0a 64 72 61 74 0a 32 36 32 31 35 33 0a 32 0a 77 74 0a 32 36 32 31
#> [3697] 35 33 0a 34 0a 71 73 65 63 0a 32 36 32 31 35 33 0a 32 0a 76 73 0a 32 36
#> [3721] 32 31 35 33 0a 32 0a 61 6d 0a 32 36 32 31 35 33 0a 34 0a 67 65 61 72 0a
#> [3745] 32 36 32 31 35 33 0a 34 0a 63 61 72 62 0a 32 35 34 0a 31 30 32 36 0a 35
#> [3769] 31 31 0a 31 36 0a 32 0a 32 36 32 31 35 33 0a 31 30 0a 61 74 74 72 69 62
#> [3793] 75 74 65 73 0a 32 36 32 31 35 33 0a 37 0a 63 6f 6c 75 6d 6e 73 0a 32 35
#> [3817] 34 0a 00 00 0b 00 00 00 e0 01 00 00 a4 01 00 00 74 01 00 00 48 01 00 00
#> [3841] 18 01 00 00 ec 00 00 00 bc 00 00 00 90 00 00 00 64 00 00 00 34 00 00 00
#> [3865] 04 00 00 00 5c fe ff ff 00 00 01 03 10 00 00 00 18 00 00 00 04 00 00 00
#> [3889] 00 00 00 00 04 00 00 00 63 61 72 62 00 00 00 00 4e fe ff ff 00 00 02 00
#> [3913] 88 fe ff ff 00 00 01 03 10 00 00 00 18 00 00 00 04 00 00 00 00 00 00 00
#> [3937] 04 00 00 00 67 65 61 72 00 00 00 00 7a fe ff ff 00 00 02 00 b4 fe ff ff
#> [3961] 00 00 01 03 10 00 00 00 14 00 00 00 04 00 00 00 00 00 00 00 02 00 00 00
#> [3985] 61 6d 00 00 a2 fe ff ff 00 00 02 00 dc fe ff ff 00 00 01 03 10 00 00 00
#> [4009] 14 00 00 00 04 00 00 00 00 00 00 00 02 00 00 00 76 73 00 00 ca fe ff ff
#> [4033] 00 00 02 00 04 ff ff ff 00 00 01 03 10 00 00 00 18 00 00 00 04 00 00 00
#> [4057] 00 00 00 00 04 00 00 00 71 73 65 63 00 00 00 00 f6 fe ff ff 00 00 02 00
#> [4081] 30 ff ff ff 00 00 01 03 10 00 00 00 14 00 00 00 04 00 00 00 00 00 00 00
#> [4105] 02 00 00 00 77 74 00 00 1e ff ff ff 00 00 02 00 58 ff ff ff 00 00 01 03
#> [4129] 10 00 00 00 18 00 00 00 04 00 00 00 00 00 00 00 04 00 00 00 64 72 61 74
#> [4153] 00 00 00 00 4a ff ff ff 00 00 02 00 84 ff ff ff 00 00 01 03 10 00 00 00
#> [4177] 14 00 00 00 04 00 00 00 00 00 00 00 02 00 00 00 68 70 00 00 72 ff ff ff
#> [4201] 00 00 02 00 ac ff ff ff 00 00 01 03 10 00 00 00 18 00 00 00 04 00 00 00
#> [4225] 00 00 00 00 04 00 00 00 64 69 73 70 00 00 00 00 9e ff ff ff 00 00 02 00
#> [4249] d8 ff ff ff 00 00 01 03 10 00 00 00 14 00 00 00 04 00 00 00 00 00 00 00
#> [4273] 03 00 00 00 63 79 6c 00 c6 ff ff ff 00 00 02 00 10 00 14 00 08 00 06 00
#> [4297] 07 00 0c 00 00 00 10 00 10 00 00 00 00 00 01 03 10 00 00 00 1c 00 00 00
#> [4321] 04 00 00 00 00 00 00 00 03 00 00 00 6d 70 67 00 00 00 06 00 08 00 06 00
#> [4345] 06 00 00 00 00 00 02 00 28 04 00 00 41 52 52 4f 57 31