Open a multi-file dataset

Arrow Datasets allow you to query against data that has been split across multiple files. This sharding of data may indicate partitioning, which can accelerate queries that only touch some partitions (files). Call open_dataset() to point to a directory of data files and return a Dataset, then use dplyr methods to query it.

open_dataset(
  sources,
  schema = NULL,
  partitioning = hive_partition(),
  unify_schemas = NULL,
  format = c("parquet", "arrow", "ipc", "feather", "csv", "tsv", "text"),
  ...
)

Arguments

sources
One of:

a string path or URI to a directory containing data files

a string path or URI to a single file

a character vector of paths or URIs to individual data files

a list of Dataset objects as created by this function

a list of DatasetFactory objects as created by dataset_factory().

When sources is a vector of file URIs, they must all use the same protocol and point to files located in the same file system and having the same format.

schema
Schema for the Dataset. If NULL (the default), the schema will be inferred from the data sources.

partitioning
When sources is a directory path/URI, one of:

a Schema, in which case the file paths relative to sources will be parsed, and path segments will be matched with the schema fields. For example, schema(year = int16(), month = int8()) would create partitions for file paths like "2019/01/file.parquet", "2019/02/file.parquet", etc.

a character vector that defines the field names corresponding to those path segments (that is, you're providing the names that would correspond to a Schema but the types will be autodetected)

a HivePartitioning or HivePartitioningFactory, as returned by hive_partition() which parses explicit or autodetected fields from Hive-style path segments

NULL for no partitioning

The default is to autodetect Hive-style partitions. When sources is not a directory path/URI, partitioning is ignored.

unify_schemas
logical: should all data fragments (files, Datasets) be scanned in order to create a unified schema from them? If FALSE, only the first fragment will be inspected for its schema. Use this fast path when you know and trust that all fragments have an identical schema. The default is FALSE when creating a dataset from a directory path/URI or vector of file paths/URIs (because there may be many files and scanning may be slow) but TRUE when sources is a list of Datasets (because there should be few Datasets in the list and their Schemas are already in memory).

format
A FileFormat object, or a string identifier of the format of the files in x. This argument is ignored when sources is a list of Dataset objects. Currently supported values:

"parquet"

"ipc"/"arrow"/"feather", all aliases for each other; for Feather, note that only version 2 files are supported

"csv"/"text", aliases for the same thing (because comma is the default delimiter for text files

"tsv", equivalent to passing format = "text", delimiter = "\t"

Default is "parquet", unless a delimiter is also specified, in which case it is assumed to be "text".

...
additional arguments passed to dataset_factory() when sources is a directory path/URI or vector of file paths/URIs, otherwise ignored. These may include format to indicate the file format, or other format-specific options.

Value

A Dataset R6 object. Use dplyr methods on it to query the data, or call $NewScan() to construct a query directly.

See also

Examples

# Set up directory for examples
tf <- tempfile()
dir.create(tf)
on.exit(unlink(tf))

data <- dplyr::group_by(mtcars, cyl)
write_dataset(data, tf)

# You can specify a directory containing the files for your dataset and
# open_dataset will scan all files in your directory.
open_dataset(tf)

#> FileSystemDataset with 3 Parquet files
#> mpg: double
#> disp: double
#> hp: double
#> drat: double
#> wt: double
#> qsec: double
#> vs: double
#> am: double
#> gear: double
#> carb: double
#> cyl: int32

# You can also supply a vector of paths
open_dataset(c(file.path(tf, "cyl=4/part-1.parquet"), file.path(tf,"cyl=8/part-2.parquet")))

#> FileSystemDataset with 2 Parquet files
#> mpg: double
#> disp: double
#> hp: double
#> drat: double
#> wt: double
#> qsec: double
#> vs: double
#> am: double
#> gear: double
#> carb: double

## You must specify the file format if using a format other than parquet.
tf2 <- tempfile()
dir.create(tf2)
on.exit(unlink(tf2))
write_dataset(data, tf2, format = "ipc")
# This line will results in errors when you try to work with the data
if (FALSE) open_dataset(tf2)
# This line will work
open_dataset(tf2, format = "ipc") 

#> FileSystemDataset with 3 Feather files
#> mpg: double
#> disp: double
#> hp: double
#> drat: double
#> wt: double
#> qsec: double
#> vs: double
#> am: double
#> gear: double
#> carb: double
#> cyl: int32

## You can specify file partitioning to include it as a field in your dataset
# Create a temporary directory and write example dataset
tf3 <- tempfile()
dir.create(tf3)
on.exit(unlink(tf3))
write_dataset(airquality, tf3, partitioning = c("Month", "Day"), hive_style = FALSE)

# View files - you can see the partitioning means that files have been written 
# to folders based on Month/Day values
list.files(tf3, recursive = TRUE)

#>   [1] "5/1/part-0.parquet"    "5/10/part-9.parquet"   "5/11/part-10.parquet" 
#>   [4] "5/12/part-12.parquet"  "5/13/part-11.parquet"  "5/14/part-13.parquet" 
#>   [7] "5/15/part-16.parquet"  "5/16/part-14.parquet"  "5/17/part-15.parquet" 
#>  [10] "5/18/part-17.parquet"  "5/19/part-18.parquet"  "5/2/part-1.parquet"   
#>  [13] "5/20/part-19.parquet"  "5/21/part-24.parquet"  "5/22/part-20.parquet" 
#>  [16] "5/23/part-25.parquet"  "5/24/part-21.parquet"  "5/25/part-26.parquet" 
#>  [19] "5/26/part-22.parquet"  "5/27/part-27.parquet"  "5/28/part-28.parquet" 
#>  [22] "5/29/part-29.parquet"  "5/3/part-2.parquet"    "5/30/part-23.parquet" 
#>  [25] "5/31/part-32.parquet"  "5/4/part-3.parquet"    "5/5/part-4.parquet"   
#>  [28] "5/6/part-5.parquet"    "5/7/part-6.parquet"    "5/8/part-7.parquet"   
#>  [31] "5/9/part-8.parquet"    "6/1/part-33.parquet"   "6/10/part-41.parquet" 
#>  [34] "6/11/part-42.parquet"  "6/12/part-43.parquet"  "6/13/part-31.parquet" 
#>  [37] "6/14/part-48.parquet"  "6/15/part-44.parquet"  "6/16/part-49.parquet" 
#>  [40] "6/17/part-50.parquet"  "6/18/part-51.parquet"  "6/19/part-52.parquet" 
#>  [43] "6/2/part-30.parquet"   "6/20/part-53.parquet"  "6/21/part-54.parquet" 
#>  [46] "6/22/part-55.parquet"  "6/23/part-56.parquet"  "6/24/part-57.parquet" 
#>  [49] "6/25/part-58.parquet"  "6/26/part-45.parquet"  "6/27/part-70.parquet" 
#>  [52] "6/28/part-46.parquet"  "6/29/part-47.parquet"  "6/3/part-34.parquet"  
#>  [55] "6/30/part-59.parquet"  "6/4/part-35.parquet"   "6/5/part-36.parquet"  
#>  [58] "6/6/part-37.parquet"   "6/7/part-38.parquet"   "6/8/part-39.parquet"  
#>  [61] "6/9/part-40.parquet"   "7/1/part-113.parquet"  "7/10/part-127.parquet"
#>  [64] "7/11/part-61.parquet"  "7/12/part-62.parquet"  "7/13/part-73.parquet" 
#>  [67] "7/14/part-74.parquet"  "7/15/part-75.parquet"  "7/16/part-90.parquet" 
#>  [70] "7/17/part-76.parquet"  "7/18/part-77.parquet"  "7/19/part-64.parquet" 
#>  [73] "7/2/part-114.parquet"  "7/20/part-65.parquet"  "7/21/part-78.parquet" 
#>  [76] "7/22/part-79.parquet"  "7/23/part-91.parquet"  "7/24/part-80.parquet" 
#>  [79] "7/25/part-81.parquet"  "7/26/part-82.parquet"  "7/27/part-63.parquet" 
#>  [82] "7/28/part-66.parquet"  "7/29/part-105.parquet" "7/3/part-71.parquet"  
#>  [85] "7/30/part-92.parquet"  "7/31/part-93.parquet"  "7/4/part-126.parquet" 
#>  [88] "7/5/part-89.parquet"   "7/6/part-72.parquet"   "7/7/part-103.parquet" 
#>  [91] "7/8/part-104.parquet"  "7/9/part-60.parquet"   "8/1/part-106.parquet" 
#>  [94] "8/10/part-99.parquet"  "8/11/part-100.parquet" "8/12/part-116.parquet"
#>  [97] "8/13/part-108.parquet" "8/14/part-117.parquet" "8/15/part-84.parquet" 
#> [100] "8/16/part-85.parquet"  "8/17/part-67.parquet"  "8/18/part-109.parquet"
#> [103] "8/19/part-118.parquet" "8/2/part-94.parquet"   "8/20/part-110.parquet"
#> [106] "8/21/part-111.parquet" "8/22/part-112.parquet" "8/23/part-86.parquet" 
#> [109] "8/24/part-87.parquet"  "8/25/part-68.parquet"  "8/26/part-119.parquet"
#> [112] "8/27/part-123.parquet" "8/28/part-101.parquet" "8/29/part-102.parquet"
#> [115] "8/3/part-83.parquet"   "8/30/part-88.parquet"  "8/31/part-69.parquet" 
#> [118] "8/4/part-115.parquet"  "8/5/part-95.parquet"   "8/6/part-107.parquet" 
#> [121] "8/7/part-96.parquet"   "8/8/part-97.parquet"   "8/9/part-98.parquet"  
#> [124] "9/1/part-124.parquet"  "9/10/part-132.parquet" "9/11/part-133.parquet"
#> [127] "9/12/part-134.parquet" "9/13/part-135.parquet" "9/14/part-136.parquet"
#> [130] "9/15/part-137.parquet" "9/16/part-138.parquet" "9/17/part-139.parquet"
#> [133] "9/18/part-140.parquet" "9/19/part-150.parquet" "9/2/part-125.parquet" 
#> [136] "9/20/part-141.parquet" "9/21/part-142.parquet" "9/22/part-143.parquet"
#> [139] "9/23/part-151.parquet" "9/24/part-144.parquet" "9/25/part-145.parquet"
#> [142] "9/26/part-146.parquet" "9/27/part-152.parquet" "9/28/part-147.parquet"
#> [145] "9/29/part-148.parquet" "9/3/part-120.parquet"  "9/30/part-149.parquet"
#> [148] "9/4/part-121.parquet"  "9/5/part-122.parquet"  "9/6/part-128.parquet" 
#> [151] "9/7/part-129.parquet"  "9/8/part-130.parquet"  "9/9/part-131.parquet" 

# With no partitioning specified, dataset contains all files but doesn't include
# directory names as field names
open_dataset(tf3)

#> FileSystemDataset with 153 Parquet files
#> Ozone: int32
#> Solar.R: int32
#> Wind: double
#> Temp: int32
#> 
#> See $metadata for additional Schema metadata

# Now that partitioning has been specified, your dataset contains columns for Month and Day
open_dataset(tf3, partitioning = c("Month", "Day"))

#> FileSystemDataset with 153 Parquet files
#> Ozone: int32
#> Solar.R: int32
#> Wind: double
#> Temp: int32
#> Month: int32
#> Day: int32
#> 
#> See $metadata for additional Schema metadata

# If you want to specify the data types for your fields, you can pass in a Schema
open_dataset(tf3, partitioning = schema(Month = int8(), Day = int8()))

#> FileSystemDataset with 153 Parquet files
#> Ozone: int32
#> Solar.R: int32
#> Wind: double
#> Temp: int32
#> Month: int8
#> Day: int8
#> 
#> See $metadata for additional Schema metadata