Read a JSON file

Using JsonTableReader

read_json_arrow(file, col_select = NULL, as_data_frame = TRUE, ...)

Arguments

file
A character file name, raw vector, or an Arrow input stream. If a file name, a memory-mapped Arrow InputStream will be opened and closed when finished; compression will be detected from the file extension and handled automatically. If an input stream is provided, it will be left open.

col_select
A character vector of column names to keep, as in the "select" argument to data.table::fread(), or a tidy selection specification of columns, as used in dplyr::select().

as_data_frame
Should the function return a data.frame (default) or an Arrow Table?

...
Additional options, passed to json_table_reader()

Value

A data.frame, or an Table if as_data_frame = FALSE.

Examples

# \donttest{
  tf <- tempfile()
  on.exit(unlink(tf))
  writeLines('
    { "hello": 3.5, "world": false, "yo": "thing" }
    { "hello": 3.25, "world": null }
    { "hello": 0.0, "world": true, "yo": null }
  ', tf, useBytes=TRUE)
  df <- read_json_arrow(tf)
# }